+++ ode/src/obstack.h
@@ -45,7 +45,7 @@ struct dObStack : public dBase {
   dObStack();
   ~dObStack();
 
-  void *alloc (int num_bytes);
+  void *alloc (size_t num_bytes);
   // allocate a block in the last arena, allocating a new arena if necessary.
   // it is a runtime error if num_bytes is larger than the arena size.
 
@@ -57,7 +57,7 @@ struct dObStack : public dBase {
   // rewind the obstack iterator, and return the address of the first
   // allocated block. return 0 if there are no allocated blocks.
 
-  void *next (int num_bytes);
+  void *next (size_t num_bytes);
   // return the address of the next allocated block. 'num_bytes' is the size
   // of the previous block. this returns null if there are no more arenas.
   // the sequence of 'num_bytes' parameters passed to next() during a
